+++ server/src/config.rs
@@ -133,7 +133,7 @@ impl TryFrom<ConfigLoad> for Config {
         let mut postgres = if let Some(postgres) = value.postgres {
             postgres
         } else {
-            return Err(XError::config("postgres is not configured").into());
+            return Err(XErr::config("postgres is not configured").into());
         };
         if postgres.application_name.is_none() {
             let application_name = USER_AGENT.to_string();