+++ server/src/services/mod.rs
@@ -1,6 +1,6 @@
-//! Services
+//! # Services
 //!
-mod api;
-mod tfb;
+mod csrf;
+mod metrics;
 
-pub use api::Api;
+pub use crate::services::{csrf::*, metrics::*};